+++ src/OpenTelemetry.Instrumentation.Http/HttpWebRequestInstrumentationOptions.netfx.cs
@@ -16,6 +16,7 @@
 
 #if NETFRAMEWORK
 using System;
+using System.Diagnostics;
 using System.Net;
 using OpenTelemetry.Context.Propagation;
 using OpenTelemetry.Instrumentation.Http.Implementation;
@@ -49,6 +50,17 @@ namespace OpenTelemetry.Instrumentation.Http
         /// If Filter returns false or throw exception, the request is filtered out.
         /// </summary>
         public Func<HttpWebRequest, bool> Filter { get; set; }
+
+        /// <summary>
+        /// Gets or sets an action to enrich an Activity.
+        /// </summary>
+        /// <remarks>
+        /// <para><see cref="Activity"/>: the activity being enriched.</para>
+        /// <para>string: the name of the event.</para>
+        /// <para>object: the raw object from which additional information can be extracted to enrich the activity.
+        /// The type of this object depends on the event, which is given by the above parameter.</para>
+        /// </remarks>
+        public Action<Activity, string, object> Enrich { get; set; }
 
         internal bool EventFilter(HttpWebRequest request)
         {