+++ src/OpenTelemetry/Trace/SamplingDecision.cs
@@ -22,21 +22,22 @@ namespace OpenTelemetry.Trace
     public enum SamplingDecision
     {
         /// <summary>
-        /// The activity object needs to be created. It will have Name, Source, Id and Baggage.
-        /// Other properties will be ignored.
+        /// The activity will be created but not recorded.
+        /// Activity.IsAllDataRequested will return false.
         /// </summary>
         Drop,
 
         /// <summary>
-        /// The activity object should be populated with all the propagation info and also all other
-        /// properties such as Links, Tags, and Events. Activity.IsAllDataRequested will return true.
+        /// The activity will be created and recorded, but sampling flag will not be set.
+        /// Activity.IsAllDataRequested will return true.
+        /// Activity.IsRecorded will return false.
         /// </summary>
         RecordOnly,
 
         /// <summary>
-        /// The activity object should be populated with all the propagation info and also all other
-        /// properties such as Links, Tags, and Events.
-        /// Both Activity.IsAllDataRequested and Activity.IsRecorded will return true.
+        /// The activity will be created, recorded, and sampling flag will be set.
+        /// Activity.IsAllDataRequested will return true.
+        /// Activity.IsRecorded will return true.
         /// </summary>
         RecordAndSample,
     }